Change return type of methods

+++ src/main/java/org/pronet/app/controllers/CommentController.java
@@ -5,6 +5,8 @@ import org.pronet.app.entities.Comment;
 import org.pronet.app.requests.comment.CommentCreateRequest;
 import org.pronet.app.requests.comment.CommentUpdateRequest;
 import org.pronet.app.services.CommentService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -17,31 +19,47 @@ public class CommentController {
     private final CommentService commentService;
 
     @PostMapping(value = "/create")
-    public Comment createComment(@RequestBody CommentCreateRequest request) {
-        return commentService.createComment(request);
+    public ResponseEntity<Comment> createComment(@RequestBody CommentCreateRequest request) {
+        Comment createdComment = commentService.createComment(request);
+        return new ResponseEntity<>(createdComment, HttpStatus.CREATED);
     }
 
     @GetMapping(value = "/list")
-    public List<Comment> getCommentList(
+    public ResponseEntity<List<Comment>> getCommentList(
             @RequestParam Optional<Long> userId,
             @RequestParam Optional<Long> postId) {
-        return commentService.getCommentList(userId, postId);
+        List<Comment> commentList = commentService.getCommentList(userId, postId);
+        return new ResponseEntity<>(commentList, HttpStatus.OK);
     }
 
     @GetMapping(value = "/details/{commentId}")
-    public Comment getCommentById(@PathVariable(value = "commentId") Long commentId) {
-        return commentService.getCommentById(commentId);
+    public ResponseEntity<Comment> getCommentById(@PathVariable(value = "commentId") Long commentId) {
+        Comment foundedComment = commentService.getCommentById(commentId);
+        if (foundedComment != null) {
+            return new ResponseEntity<>(foundedComment, HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @PutMapping(value = "/update/{commentId}")
-    public Comment updateComment(
+    public ResponseEntity<Comment> updateComment(
             @PathVariable(value = "commentId") Long commentId,
             @RequestBody CommentUpdateRequest request) {
-        return commentService.updateComment(commentId, request);
+        Boolean existCommentById = commentService.existsById(commentId);
+        if (existCommentById) {
+            Comment updatedComment = commentService.updateComment(commentId, request);
+            return new ResponseEntity<>(updatedComment, HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @DeleteMapping(value = "/delete/{commentId}")
-    public void deleteComment(@PathVariable(value = "commentId") Long commentId) {
-        commentService.deleteComment(commentId);
+    public ResponseEntity<Void> deleteComment(@PathVariable(value = "commentId") Long commentId) {
+        Boolean existCommentById = commentService.existsById(commentId);
+        if (existCommentById) {
+            commentService.deleteComment(commentId);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 }
